Left sidebar separated and rearranged. Second sidebar renamed and rearranged
