+++ client/src/components/ChatBox/CurrentChatWindow/UserNav/index.js
@@ -1,11 +1,9 @@
-import {useContext} from 'react'
-import { useHistory } from 'react-router-dom'
+import { useContext } from 'react'
 import styles from './index.module.css'
 import { AuthenticateUser } from '../../../../context/AuthenticationContext'
 import { MessagesContext } from '../../../../context/MessagesContext'
 
 const UserNav = () => {
-    const history = useHistory()
     const { logOut } = AuthenticateUser()
     const { dispatchUserData } = useContext(MessagesContext)
 
@@ -22,7 +20,6 @@ const UserNav = () => {
             <button onClick={loadProjects}>Projects</button>
             <button onClick={loadProfile}>Profile</button>
             <button onClick={logOut}>Logout</button>
-            <button onClick={() => history.push('/chat')}>Old</button>
         </div>
     )
 }
