+++ client/src/components/CommonComponents/NewMessageLight/index.js
@@ -1,0 +1,9 @@
+import styles from './index.module.css'
+
+const NewMessageLight = () => {
+    return (
+        <div className={styles['new-message']}></div>
+    )
+}
+
+export default NewMessageLight
