--- client/src/components/CommonComponents/TransparentBackground/index.module.css
@@ -1,11 +1,0 @@
-.window {
-    position: fixed;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    width: 100%;
-    height: 100%;
-    background-color: black;
-    z-index: 1;
-    opacity: 50%;
-}
