+++ client/src/components/ProjectSidebar/GroupsList/index.module.css
@@ -1,0 +1,4 @@
+.container {
+    display: flex;
+    flex-direction: column;
+}
