+++ client/src/components/ProjectSidebar/Member/index.js
@@ -1,0 +1,40 @@
+import { useContext } from 'react'
+import styles from './index.module.css'
+import StatusLight from '../../CommonComponents/StatusLight'
+import UserAvatar from '../../CommonComponents/UserAvatar'
+import { MessagesContext } from '../../../context/MessagesContext'
+import { SocketContext } from '../../../context/SocketContext'
+import infoButton from '../../../images/informationButton.svg'
+
+const Member = ({ id, name, picturePath, isOnline }) => {
+    const { dispatchUserData } = useContext(MessagesContext)
+    const { socket } = useContext(SocketContext)
+
+    const openPrivateChat = () => {
+        socket.emit('get-chat-history', id, (chat) => {
+            dispatchUserData({ type: 'open-chat', payload: { id, chat } })
+        })
+    }
+
+    return (
+        <div className={styles.container}>
+            <div className={styles.member} onClick={openPrivateChat}>
+                <div className={styles.status}>
+                    <StatusLight userId={id} isOnline={isOnline} size='small' />
+                </div>
+                <UserAvatar picturePath={picturePath} />
+                <div className={styles.name}>
+                    {name}
+                </div>
+            </div>
+            <img 
+                src={infoButton} 
+                className={styles.info} 
+                alt='Link to user info'
+                onClick={() => dispatchUserData({ type: 'show-details', id, show: true })}
+            />
+        </div>
+    )
+}
+
+export default Member
