+++ client/src/components/ProjectSidebar/Member/index.module.css
@@ -1,0 +1,29 @@
+.container {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    margin-bottom: 11px;
+    width: 100%;
+}
+
+.member {
+    display: flex;
+}
+
+.status {
+    display: flex;
+    align-items: center;
+}
+
+.name {
+    margin-left: 10px;
+    font-size: 14px;
+    color: #B5B5B5;
+    display: flex;
+    align-items: center;
+}
+
+.info {
+    max-width: 1.2rem;
+}
+
