+++ client/src/components/ProjectSidebar/MembersList/index.module.css
@@ -1,0 +1,15 @@
+.container {
+    display: flex;
+    flex-direction: column;
+    flex-basis: 225px;
+    margin-top: 24px;
+    padding: 0 1rem 0 0.5rem;
+}
+
+.header {
+    font-size: 12px;
+    font-weight: bold;
+    color: #919191;
+    display: flex;
+    justify-content: center;
+}
