+++ client/src/components/ProjectSidebar/index.module.css
@@ -1,0 +1,10 @@
+.container {
+    background-color: #707070;
+    display: flex; 
+    flex-direction: column;
+    flex-basis: 225px;
+    flex-grow: 0;
+    flex-shrink: 0;
+    overflow-y: auto;
+    scrollbar-width: none;
+}
