--- client/src/components/ProjectsListBox/NewMessageLight/index.js
@@ -1,10 +1,0 @@
-import React from 'react'
-import styles from './index.module.css'
-
-const NewMessageLight = () => {
-    return (
-        <div className={styles['new-message']}></div>
-    )
-}
-
-export default NewMessageLight
