--- client/src/components/ProjectsListBox/NewProject/index.module.css
@@ -1,26 +1,0 @@
-.button {
-    height: 40px;
-    width: 40px;
-    background-color: #5D5D5D;
-    border-radius: 50%; 
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    margin-bottom: 10px;
-}
-
-.img {
-    display: inline-block; 
-    
-}
-
-.box {
-    display: flex;
-    flex-direction: column;
-    width: 70px;
-    flex-basis: auto;
-    align-items: center;
-    justify-content: center;
-    margin-bottom: 15px;
-    margin-top: 15px;
-}
