--- client/src/components/ProjectsListBox/ProjectCircle/index.module.css
@@ -1,4 +1,0 @@
-.circle {
-   margin-left: 5px;
-   margin-bottom: 5px;
-}
