--- client/src/components/ProjectsListBox/StatusLightsBox/index.module.css
@@ -1,6 +1,0 @@
-.status-lights-box {
-    display: flex;
-    height: 30px;
-    flex: 1 0 40px;
-    flex-grow: 0;
-}
