--- client/src/components/SelectedProjectBox/Friend/index.module.css
@@ -1,28 +1,0 @@
-.friends {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    margin-bottom: 11px;
-    width: 100%;
-}
-
-.status-light {
-    display: flex;
-    align-items: center;
-}
-
-.name {
-    margin-left: 10px;
-    font-size: 14px;
-    color: #B5B5B5;
-    display: flex;
-    align-items: center;
-}
-
-.btn-info {
-    max-width: 1.2rem;
-}
-
-.friend-main {
-    display: flex;
-}
