--- client/src/components/SelectedProjectBox/FriendsBox/index.module.css
@@ -1,5 +1,0 @@
-.friends-box {
-    display: flex;
-    flex-basis: 225px;
-    flex-direction: column;
-}
