--- client/src/components/SelectedProjectBox/FriendsList/index.module.css
@@ -1,10 +1,0 @@
-.friends-list {
-    display: flex;
-    flex-direction: column;
-    margin-top: 24px;
-    padding: 0 1rem 0 0.5rem;
-}
-
-ul li {
-    padding: 0;
-}
