--- client/src/components/SelectedProjectBox/FriendsNumber/index.module.css
@@ -1,19 +1,0 @@
-.friends {
-    
-    margin-top: 30px;
-    font-size: 12px;
-    font-weight: bold;
-    color: #919191;
-    display: flex;
-    /* flex-basis: 225px; */
-    
-}
-
-.friends-label{
-    margin-left: 25px;
-}
-
-.number {
-    margin-left: auto;
-    margin-right: 10px;
-}
