--- client/src/components/SelectedProjectBox/ProjectList/index.module.css
@@ -1,37 +1,0 @@
-.project-list {
-    color: #FFFFFF;
-    font-size: 20px;
-    margin-left: 24px;
-    margin-top: 40px;
-    display: flex;
-    align-items: center;
-}
-
-.plus-icon {
-    width: 14px;
-    height: 14px;
-    margin-right: 5px;
-    margin-left: auto;
-}
-
-.settings-icon {
-    width: 14px;
-    height: 14px;
-    margin-right: 15px;
-}
-
-.small-icons{
-    margin-left: auto;
-}
-
-.info {
-    margin-right: 15px;
-    border: 1px solid white;
-    border-radius: 50%;
-    text-align: center;
-    width: 16px;
-    height: 16px;
-    line-height: 16px;
-    color: white;
-    font-size: x-small;
-}
