--- client/src/components/SelectedProjectBox/ProjectTreads/Thread/index.module.css
@@ -1,24 +1,0 @@
-.tread {
-    width: 150px;
-    height: 30px;
-    border: 2px; 
-    margin-bottom: 10px;
-    font-size: 14px;
-    margin-left: 10px;
-    padding: 10px;
-    color: #B5B5B5;
-    flex-grow: 0;
-}
-
-.tread:hover {
-    width: 150px;
-    height: 30px;
-    border: 2px; 
-    margin-bottom: 10px;
-    font-size: 14px;
-    margin-left: 10px;
-    padding: 10px;
-    color: white;
-    background-color: #7C7D7F;
-    border-radius: 5px;
-}
