--- client/src/components/SelectedProjectBox/ProjectTreads/index.module.css
@@ -1,37 +1,0 @@
-.project-treads {
-    display: flex;
-    flex-direction: column;
-    border: 2px;
-    flex-grow: 0;
-    list-style-type: none;
-}
-
-.tread {
-    width: 150px;
-    height: 30px;
-    border: 2px; 
-    margin-bottom: 10px;
-    font-size: 14px;
-    margin-left: 10px;
-    padding: 10px;
-    color: #B5B5B5;
-    flex-grow: 0;
-}
-
-.tread:hover {
-    width: 150px;
-    height: 30px;
-    border: 2px; 
-    margin-bottom: 10px;
-    font-size: 14px;
-    margin-left: 10px;
-    padding: 10px;
-    color: white;
-    background-color: #7C7D7F;
-    border-radius: 5px;
-}
-
-.list {
-    margin: 0px;
-    padding: 0px;
-}
