+++ client/src/pages/MainPage/index.js
@@ -2,8 +2,8 @@ import React from 'react'
 import styles from './index.module.css'
 import SocketContextProvider from '../../context/SocketContext'
 import MessagesContextProvider from '../../context/MessagesContext'
-import ProjectsList from '../../components/ProjectsListBox/ProjectsList'
-import SelectedProject from '../../components/SelectedProjectBox/SelectedProject'
+import LeftSidebar from '../../components/LeftSidebar'
+import ProjectSidebar from '../../components/ProjectSidebar'
 import ChatBox from '../../components/ChatBox'
 import ProfileInfoBox from '../../components/ProfileInfoBox'
 
@@ -12,8 +12,8 @@ const ChatPageNewDesign = () => {
         <SocketContextProvider>
             <MessagesContextProvider>
                 <div className={styles['chat-page-new-design']}>
-                    <ProjectsList />
-                    <SelectedProject />
+                    <LeftSidebar />
+                    <ProjectSidebar />
                     <ChatBox />
                     <ProfileInfoBox />
                 </div>
